git-svn-id: https://svn.code.sf.net/p/web-erp/code/trunk@5466 09516a42-71f9-43be-811f-9b821cf0d246

+++ includes/Z_POSDataCreation.php
@@ -1,6 +1,6 @@
 <?php
 
-function CreatePOSDataFull ( $POSDebtorNo, $POSBranchCode, $db) {
+function Create_POS_Data_Full ( $POSDebtorNo, $POSBranchCode, $db) {
 	$result = DB_query("SELECT currcode, salestype FROM debtorsmaster WHERE debtorno='" . $POSDebtorNo . "'",$db);
 	$CustomerRow = DB_fetch_array($result);
 	if (DB_num_rows($result)==0){
@@ -11,11 +11,15 @@ function CreatePOSDataFull ( $POSDebtorNo, $POSBranchCode, $db) {
 
 	$FileHandle = fopen($_SESSION['reports_dir'] . '/POS.sql','w');
 
+	if ($FileHandle == false){
+		return 0;
+	}
+
 	fwrite($FileHandle,"DELETE FROM currencies;\n");
 	$result = DB_query('SELECT currency, currabrev, country, hundredsname,decimalplaces, rate FROM currencies',$db);
 	while ($CurrRow = DB_fetch_array($result)) {
 
-		  fwrite($FileHandle,"INSERT INTO currencies VALUES ('" . $CurrRow['currency'] . "', '" . $CurrRow['currabrev'] . "', '" . sqlite_escape_string ($CurrRow['country']) . "', '" . sqlite_escape_string ($CurrRow['hundredsname']) . "', '" .$CurrRow['decimalplaces'] . "', '" .$CurrRow['rate'] . "');\n");
+		  fwrite($FileHandle,"INSERT INTO currencies VALUES ('" . $CurrRow['currency'] . "', '" . $CurrRow['currabrev'] . "', '" . SQLite_Escape ($CurrRow['country']) . "', '" . SQLite_Escape ($CurrRow['hundredsname']) . "', '" .$CurrRow['decimalplaces'] . "', '" .$CurrRow['rate'] . "');\n");
 
 	}
 
@@ -24,7 +28,7 @@ function CreatePOSDataFull ( $POSDebtorNo, $POSBranchCode, $db) {
 	$result = DB_query('SELECT typeabbrev, sales_type FROM salestypes',$db);
 	while ($myrow = DB_fetch_array($result)) {
 
-		  fwrite($FileHandle,"INSERT INTO salestypes VALUES ('" . $myrow['typeabbrev'] . "', '" . sqlite_escape_string ($myrow['sales_type']) . "');\n");
+		  fwrite($FileHandle,"INSERT INTO salestypes VALUES ('" . $myrow['typeabbrev'] . "', '" . SQLite_Escape ($myrow['sales_type']) . "');\n");
 
 	}
 	fwrite($FileHandle,"DELETE FROM holdreasons;\n");
@@ -32,7 +36,7 @@ function CreatePOSDataFull ( $POSDebtorNo, $POSBranchCode, $db) {
 	$result = DB_query('SELECT reasoncode, reasondescription, dissallowinvoices FROM holdreasons',$db);
 	while ($myrow = DB_fetch_array($result)) {
 
-		  fwrite($FileHandle,"INSERT INTO holdreasons VALUES ('" . $myrow['reasoncode'] . "', '" . sqlite_escape_string ($myrow['reasondescription']) . "', '" . $myrow['dissallowinvoices'] . "');\n");
+		  fwrite($FileHandle,"INSERT INTO holdreasons VALUES ('" . $myrow['reasoncode'] . "', '" . SQLite_Escape ($myrow['reasondescription']) . "', '" . $myrow['dissallowinvoices'] . "');\n");
 
 	}
 	fwrite($FileHandle,"DELETE FROM paymentterms;\n");
@@ -40,7 +44,7 @@ function CreatePOSDataFull ( $POSDebtorNo, $POSBranchCode, $db) {
 	$result = DB_query('SELECT termsindicator, terms FROM paymentterms',$db);
 	while ($myrow = DB_fetch_array($result)) {
 
-		  fwrite($FileHandle,"INSERT INTO paymentterms VALUES ('" . $myrow['termsindicator'] . "', '" . sqlite_escape_string ($myrow['terms']) . "');\n");
+		  fwrite($FileHandle,"INSERT INTO paymentterms VALUES ('" . $myrow['termsindicator'] . "', '" . SQLite_Escape ($myrow['terms']) . "');\n");
 
 	}
 
@@ -48,7 +52,7 @@ function CreatePOSDataFull ( $POSDebtorNo, $POSBranchCode, $db) {
 	$result = DB_query('SELECT paymentid, paymentname,opencashdrawer FROM paymentmethods',$db);
 	while ($myrow = DB_fetch_array($result)) {
 
-		  fwrite($FileHandle,"INSERT INTO paymentmethods VALUES ('" . $myrow['paymentid'] . "', '" . sqlite_escape_string ($myrow['paymentname']) . "', '" . $myrow['opencashdrawer'] . "');\n");
+		  fwrite($FileHandle,"INSERT INTO paymentmethods VALUES ('" . $myrow['paymentid'] . "', '" . SQLite_Escape ($myrow['paymentname']) . "', '" . $myrow['opencashdrawer'] . "');\n");
 
 	}
 
@@ -56,7 +60,7 @@ function CreatePOSDataFull ( $POSDebtorNo, $POSBranchCode, $db) {
 	$result = DB_query('SELECT loccode, locationname,taxprovinceid FROM locations',$db);
 	while ($myrow = DB_fetch_array($result)) {
 
-		  fwrite($FileHandle,"INSERT INTO locations VALUES ('" . $myrow['loccode'] . "', '" . sqlite_escape_string ($myrow['locationname']) . "', '" . $myrow['taxprovinceid'] . "');\n");
+		  fwrite($FileHandle,"INSERT INTO locations VALUES ('" . $myrow['loccode'] . "', '" . SQLite_Escape ($myrow['locationname']) . "', '" . $myrow['taxprovinceid'] . "');\n");
 
 	}
 
@@ -64,14 +68,14 @@ function CreatePOSDataFull ( $POSDebtorNo, $POSBranchCode, $db) {
 	$result = DB_query('SELECT categoryid, categorydescription FROM stockcategory',$db);
 	while ($myrow = DB_fetch_array($result)) {
 
-		  fwrite($FileHandle,"INSERT INTO stockcategory VALUES ('" . $myrow['categoryid'] . "', '" . sqlite_escape_string ($myrow['categorydescription']) . "');\n");
+		  fwrite($FileHandle,"INSERT INTO stockcategory VALUES ('" . $myrow['categoryid'] . "', '" . SQLite_Escape ($myrow['categorydescription']) . "');\n");
 
 	}
 	fwrite($FileHandle,"DELETE FROM taxgroups;\n");
 	$result = DB_query('SELECT taxgroupid, taxgroupdescription FROM taxgroups',$db);
 	while ($myrow = DB_fetch_array($result)) {
 
-		  fwrite($FileHandle,"INSERT INTO taxgroups VALUES ('" . $myrow['taxgroupid'] . "', '" . sqlite_escape_string ($myrow['taxgroupdescription']) . "');\n");
+		  fwrite($FileHandle,"INSERT INTO taxgroups VALUES ('" . $myrow['taxgroupid'] . "', '" . SQLite_Escape ($myrow['taxgroupdescription']) . "');\n");
 
 	}
 
@@ -86,7 +90,7 @@ function CreatePOSDataFull ( $POSDebtorNo, $POSBranchCode, $db) {
 	$result = DB_query('SELECT taxid, description FROM taxauthorities',$db);
 	while ($myrow = DB_fetch_array($result)) {
 
-		  fwrite($FileHandle,"INSERT INTO taxauthorities VALUES ('" . $myrow['taxid'] . "', '" . sqlite_escape_string ($myrow['description']) . "');\n");
+		  fwrite($FileHandle,"INSERT INTO taxauthorities VALUES ('" . $myrow['taxid'] . "', '" . SQLite_Escape ($myrow['description']) . "');\n");
 
 	}
 	fwrite($FileHandle,"DELETE FROM taxauthrates;\n");
@@ -100,8 +104,8 @@ function CreatePOSDataFull ( $POSDebtorNo, $POSBranchCode, $db) {
 	$result = DB_query("SELECT stockid, categoryid, description, longdescription, units, barcode, taxcatid, decimalplaces FROM stockmaster WHERE (mbflag='B' OR mbflag='M') AND discontinued=0 AND controlled=0",$db);
 	while ($myrow = DB_fetch_array($result)) {
 
-		  fwrite($FileHandle,"INSERT INTO stockmaster VALUES ('" . sqlite_escape_string ($myrow['stockid']) . "', '" . sqlite_escape_string ($myrow['categoryid']) . "', '" . sqlite_escape_string ($myrow['description']) . "', '" . sqlite_escape_string (str_replace("\n", '', $myrow['longdescription'])) . "', '" . sqlite_escape_string ($myrow['units']) . "', '" . sqlite_escape_string ($myrow['barcode']) . "', '" . $myrow['taxcatid'] . "', '" . $myrow['decimalplaces'] . "');\n");
-	      $Price = GetPrice ($myrow['stockid'], $_POST['POSDebtorNo'], $_POST['POSBranchCode'], $db,0);
+		  fwrite($FileHandle,"INSERT INTO stockmaster VALUES ('" . SQLite_Escape ($myrow['stockid']) . "', '" . SQLite_Escape ($myrow['categoryid']) . "', '" . SQLite_Escape ($myrow['description']) . "', '" . SQLite_Escape (str_replace("\n", '', $myrow['longdescription'])) . "', '" . SQLite_Escape ($myrow['units']) . "', '" . SQLite_Escape ($myrow['barcode']) . "', '" . $myrow['taxcatid'] . "', '" . $myrow['decimalplaces'] . "');\n");
+	      $Price = GetPriceQuick ($myrow['stockid'], $_POST['POSDebtorNo'], $_POST['POSBranchCode'], $db);
 	      if ($Price!=0) {
 		  	  fwrite($FileHandle,"INSERT INTO prices (stockid, currabrev, typeabbrev, price) VALUES('" . $myrow['stockid'] . "', '" . $CurrCode . "', '" . $SalesType . "', '" . $Price . "');\n");
   		  }
@@ -111,14 +115,14 @@ function CreatePOSDataFull ( $POSDebtorNo, $POSBranchCode, $db) {
 	$result = DB_query("SELECT debtorno, name, currcode, salestype, holdreason, paymentterms, discount, creditlimit, discountcode FROM debtorsmaster WHERE currcode='". $CurrCode . "'",$db);
 	while ($myrow = DB_fetch_array($result)) {
 
-		  fwrite($FileHandle,"INSERT INTO debtorsmaster VALUES ('" . $myrow['debtorno'] . "', '" . sqlite_escape_string ($myrow['name']) . "', '" . $myrow['currcode'] . "', '" . $myrow['salestype'] . "', '" . $myrow['holdreason'] . "', '" . sqlite_escape_string ($myrow['paymentterms']) . "', '" . $myrow['discount'] . "', '" . $myrow['creditlimit'] . "', '" . $myrow['discountcode'] . "');\n");
+		  fwrite($FileHandle,"INSERT INTO debtorsmaster VALUES ('" . $myrow['debtorno'] . "', '" . SQLite_Escape ($myrow['name']) . "', '" . $myrow['currcode'] . "', '" . $myrow['salestype'] . "', '" . $myrow['holdreason'] . "', '" . SQLite_Escape ($myrow['paymentterms']) . "', '" . $myrow['discount'] . "', '" . $myrow['creditlimit'] . "', '" . $myrow['discountcode'] . "');\n");
 
 	}
 	fwrite($FileHandle,"DELETE FROM custbranch;\n");
 	$result = DB_query("SELECT branchcode, debtorsmaster.debtorno, brname, contactname, specialinstructions,taxgroupid FROM custbranch INNER JOIN debtorsmaster ON custbranch.debtorno=debtorsmaster.debtorno WHERE debtorsmaster.currcode='". $CurrCode . "'",$db);
 	while ($myrow = DB_fetch_array($result)) {
 
-		  fwrite($FileHandle,"INSERT INTO custbranch VALUES ('" . $myrow['branchcode'] . "', '" . $myrow['debtorno'] . "', '" . sqlite_escape_string ($myrow['brname']) . "', '" . sqlite_escape_string ($myrow['contactname']) . "', '" . sqlite_escape_string ($myrow['specialinstructions']) . "', '" . $myrow['taxgroupid'] . "');\n");
+		  fwrite($FileHandle,"INSERT INTO custbranch VALUES ('" . $myrow['branchcode'] . "', '" . $myrow['debtorno'] . "', '" . SQLite_Escape ($myrow['brname']) . "', '" . SQLite_Escape ($myrow['contactname']) . "', '" . SQLite_Escape ($myrow['specialinstructions']) . "', '" . $myrow['taxgroupid'] . "');\n");
 
 	}
 
@@ -134,6 +138,25 @@ function CreatePOSDataFull ( $POSDebtorNo, $POSBranchCode, $db) {
 	}
 	$ZipFile->addFile($_SESSION['reports_dir'] . '/POS.sql','POS.sql');
 	$ZipFile->close();
+	//delete the original big sql file as we now have the zip for transferring
+	unlink($_SESSION['reports_dir'] . '/POS.sql');
+	return 1;
 }
 
+function SQLite_Escape($String) {
+  $SearchCharacters  = array('&', '"', "'",'<', '>',"\n","\r"    );
+  $ReplaceWith = array('&amp;', '""', "''", '&lt;', '&gt;', '', '&#13;');
+
+  $String = str_replace($SearchCharacters, $ReplaceWith, $String);
+  return $String;
+}
+function Delete_POS_Data(){
+	if (file_exists($_SESSION['reports_dir'] . '/POS.sql.zip')){
+		unlink($_SESSION['reports_dir'] . '/POS.sql');
+	}
+	if (file_exists($_SESSION['reports_dir'] . '/POS.sql')){
+		unlink($_SESSION['reports_dir'] . '/POS.sql');
+	}
+	return 1;
+}
 ?>
